+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 * Tools: fixed a crash when re-importing resources in some circumnstances.
 * Tools: fixed flythrough camera excessive speed when moving diagonally.
 * Tools: fixed re-importing existing units sometimes losing component data.
+* Tools: fixed imported components not being removed from units when disabled on re-import.
 * Docs: updated Android deploying instructions.
 
 .. _v0.63.1: